# senate-portal/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreateAttendanceRequestParams": {
        "type": "object",
        "properties": {
          "attendanceId": {
            "type": "string",
            "description": "Attendance Id"
          },
          "reason": {
            "type": "string",
            "description": "Reason for the request"
          },
          "attendanceMode": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AttendanceMode"
              }
            ],
            "description": "Attendance Mode, online or in-person"
          },
          "TimeAdjustment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeAdjustment"
              }
            ],
            "description": "Time Adjustment, arriving late or leaving early"
          }
        },
        "required": [
          "attendanceId",
          "reason",
          "attendanceMode",
          "TimeAdjustment"
        ]
      },
      "CreateAttendanceRequestResponse": {
        "type": "object",
        "properties": {
          "attendanceId": {
            "type": "string",
            "description": "Attendance Id"
          },
          "requestId": {
            "type": "string",
            "description": "Request Id"
          },
          "reason": {
            "type": "string",
            "description": "Reason for the request"
          },
          "attendanceMode": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AttendanceMode"
              }
            ],
            "description": "Attendance Mode, online or in-person"
          },
          "TimeAdjustment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeAdjustment"
              }
            ],
            "description": "Time Adjustment, arriving late or leaving early"
          },
          "isLate": {
            "type": "boolean",
            "description": "Is Late"
          }
        },
        "required": [
          "attendanceId",
          "requestId",
          "reason",
          "attendanceMode",
          "TimeAdjustment",
          "isLate"
        ]
      },
...
      "AttendanceResponse": {
        "type": "object",
        "properties": {
          "attendanceId": {
            "type": "string",
            "description": "Attendance Id"
          },
          "userId": {
            "type": "string",
            "description": "User Id"
          },
          "meetingId": {
            "type": "string",
            "description": "Meeting Id"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AttendanceStatus"
              }
            ],
            "description": "Status of the attendance"
          }
        },
        "required": ["attendanceId", "userId", "meetingId", "status"]
      },
...
      "PostAttendanceParams": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User Id"
          },
          "meetingId": {
            "type": "string",
            "description": "Meeting Id"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AttendanceStatus"
              }
            ],
            "description": "Status of the attendance"
          }
        },
        "required": ["userId", "meetingId", "status"]
      },
...
      "AttendanceStatus": {
        "type": "string",
        "enum": ["PRESENT", "UNEXCUSED_ABSENCE", "EXCUSED_ABSENCE", "PENDING"]
      },
      "AttendanceMode": {
        "type": "string",
        "enum": ["ONLINE", "IN_PERSON"]
      },
      "TimeAdjustment": {
        "type": "string",
        "enum": ["ARRIVING_LATE", "LEAVING_EARLY"]
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/attendance": {
      "post": {
        "operationId": "post-attendance",
        "summary": "Posts an Attendance",
        "description": "Creates a new Attendance Record with params",
        "tags": ["Attendance"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostAttendanceParams"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/attendance/{attendanceId}/requests": {
      "post": {
        "operationId": "post-attendance-{attendanceId}-requests",
        "summary": "Create a new Request",
        "description": "Creates a new Request for specific attendanceId",
        "tags": ["Attendance"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttendanceRequestParams"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAttendanceRequestResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/requests": {
      "get": {
        "operationId": "get-requests",
        "summary": "GET all requests (for admins)",
        "description": "",
        "tags": ["Requests"],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-requests",
        "summary": "POST create new request",
        "description": "",
        "tags": ["Requests"],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/requests/{id}": {
      "get": {
        "operationId": "get-requests-{id}",
        "summary": "",
        "description": "",
        "tags": ["Requests"],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "operationId": "put-requests-{id}",
        "summary": "",
        "description": "",
        "tags": ["Requests"],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "operationId": "delete-requests-{id}",
        "summary": "",
        "description": "",
        "tags": ["Requests"],
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}